+++ src-webviews/src/pages/actions/interfaces/IAction.ts
@@ -1,0 +1,5 @@
+export default interface IAction {
+    eventName: string;
+    args: Array<unknown>;
+    isServer: boolean;
+}